+++ root/programs/CS/Frameworks/Infrastructure/Framework/Transmission/ProtocolNameService.cs
@@ -91,7 +91,7 @@ namespace Touryo.Infrastructure.Framework.Transmission
             {
                 // 呼び出しプロトコルの名前解決定義（XmlDocument）を[リソース]で初期化
                 this.XMLTMD_Protocol.Load(
-                    PubCmnFunction.BuiltStringIntoEnvironmentVariable(
+                    StringVariableOperator.BuiltStringIntoEnvironmentVariable(
                         GetConfigParameter.GetConfigValue(FxLiteral.XML_TM_PROTOCOL_DEFINITION)));
 
                 // 戻す
@@ -403,7 +403,7 @@ namespace Touryo.Infrastructure.Framework.Transmission
 
                             // プロパティ文字列をディクショナリに変換
                             string propStr = xmlNode.Value;
-                            props = PubCmnFunction.GetPropsFromPropString(propStr);
+                            props = StringVariableOperator.GetPropsFromPropString(propStr);
                         }
                     }
                 }
